pull search type/arg from form fields again.
Reverted to using the form fields for type an argument to allow for a button href update whenever there was a change but not a page submit. Also allowed for better handling of special characters, spaces, etc. The button is now on the bibdisplay page - it grabs the title from the page - throws away anything after a forward slash, and strips out any non alphanumeric characters from what's left. Then it wraps that in quotes and submits it to prospector as a phrase keyword search. This commit also includes some improvements to the sematics of various markup and minor css changes.